Add files via upload
This is an project on pharma classification of liver diseases so there are four different types of  liver diseases. I have developed many types of models and check for best accuracy among them XGBOOST model gave best accuracy. I finalized it as the best model and deceloped a .py file for deployment .